requirements.txt
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
decorator==5.1.1
Django==4.2
django-filter==24.3
django-stubs==5.1.2
django-stubs-ext==5.1.2
djangorestframework==3.15.2
executing==2.2.0
flower==2.0.1
gunicorn==23.0.0
humanize==4.11.0
idna==3.10
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.32.0
jedi==0.19.2
kombu==5.4.2
matplotlib-inline==0.1.7
mypy==1.14.1
mypy-extensions==1.0.0
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
types-PyYAML==6.0.12.20241230
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
